🔧 update (yak-map): split render.rs into render/ module with task.rs
Broke 645-LOC render.rs into render/mod.rs (highlight_line + 3 tests)
and render/task.rs (task rendering + 33 tests). All 70 tests pass,
WASM builds, no API changes.
